[Feature] Implemented placeholder screens
Need to refine these for functionality and to look more like storyboard/prototype, just wanted to have something down so that we have a skeleton for a few main screens and navigation between them. Experimenting with ReactNative elements, this styling should be baked into a separate file instead of tagged onto the end of our pages.