test I was working on
before keith hit legacy conflict, and wanted to pause to communicate how to resolve those. I wanted to get this in here to after i scrapped my dependency needed so i didnt overcommit into this branch as we mentioned wanting to work in the same branch to do that at one point. This way I can show I was working on a test case.